plataforma_automacao_ia_generativa/mvp_concierge/backend_concierge_scripts/requirements.txt
google-generativeai
python-dotenv
reportlab
uvicorn
fastapi